requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
google-cloud-bigquery==3.21.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2